♻️  Front: extract getFullName from UserService to its own utils file
